make project multifile
